refactor: 남은시간 계산 함수 분리
